move disposeasync to gridreader.async
